pyproject.toml
[tool.poetry]
name = "AVA"
version = "0.0.1"
description = "Smart and intelligent virtual assistant with ability to upload your own API token for more personalized and secured conversation."
authors = ["Jason Kim <support@onebladesolutions.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
pydantic = "^1.10.7"
pyhumps = "^3.8.0"
uvicorn = "^0.22.0"
mangum = "^0.17.0"
dynamodb-json = "^1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
Jinja2 = "^3.1.2"
cachetools = "^5.3.0"
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
requests = "^2.30.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
moto = {extras = ["dynamodb"], version = "^4.1.9"}

[tool.poetry.group.dev.dependencies]
boto3-stubs = {extras = ["essential", "secretsmanager"], version = "^1.26.137"}
boto3 = "^1.26.137"
types-boto3 = "^1.0.2"
types-passlib = "^1.7.7.12"
types-python-jose = "^3.3.4.7"
types-cachetools = "^5.3.0.5"
types-requests = "^2.30.0.0"
fuzzywuzzy-stubs = "^0.0.1"
black = "^23.3.0"
pylint = "^2.17.4"
mypy = "^1.3.0"
bumpversion = "^0.6.0"
types-python-dateutil = "^2.8.19.13"
types-pytz = "^2023.3.0.0"
pytest = "^7.3.1"
freezegun = "^1.2.2"
pytest-cov = "^4.0.0"
pre-commit = "^3.3.1"
ruff = "^0.0.265"
isort = "^5.12.0"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
extend-exclude = ["deploy/*", "tests/*"]
line-length = 120
target-version = "py310"
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"